fixes minor float bug

+++ NMT/src/trainer.py
@@ -437,7 +437,7 @@ class TrainerMT(MultiprocessingEventLoop):
 						mul1 = src_labels.unsqueeze(-1) 
 						mul2 = (src_labels - 1.0).unsqueeze(-1)
 
-						loss1 += -1.0 (torch.mm(mul1.t(), prediction) / (1.0 + torch.sum(src_labels)) + torch.mm(mul2.t(), prediction) / (1.0 + torch.sum(1.0 - src_labels)))
+						loss1 += -1.0 (torch.mm(mul1.float().t(), prediction) / (1.0 + torch.sum(src_labels)) + torch.mm(mul2.float().t(), prediction) / (1.0 + torch.sum(1.0 - src_labels)))
 					else:
 						loss1 += F.cross_entropy(prediction,src_labels)
 			self.stats['seq_dis_costs'].append(loss1.item())
